# Prevent Copying Schema on every Write + Read