@akrennmair 2022-01-29
Hi, sorry for not getting back to you earlier. IIRC, there was a specific reason why we implemented a deep copy of the `GetSchemaDefinition` method like that, but I can't remember it. I found the revision where this was changed, `c5d735c1`, and all it says is "fix the schema copy issue". Maybe @fzerorubigd remembers more about this?