@deankarn 2022-01-29
@akrennmair I can’t see a reason beyond the schema changing mid-flight.

either way let me know and if it is indeed required I can create a PR to make the deep copy more efficient.