@fzerorubigd 2022-01-29
There was a use case to create a copy of a parquet file with different schema, (adding or deleting a field) the problem was by modifying the target schema the original one was modified therefore we implemented the deep copy.