@deankarn 2022-01-29
@fzerorubigd this sounds like a special use case, copying a file, which makes sense but also a misuse of the same schema and not needed during normal operation; copying the schema on each read +write.

It seems this may have slipped into the normal operation because the function in question wasn’t clear on what it did.

I propose keeping the function in this PR returning the already parsed schema, as it’s named, and add a new function denoting a copy eg. Clone()

This way you can have the functionality for both use cases and it’s usage is clear.

let me know what you think :)