@akrennmair 2022-01-29
@deankarn that sounds like a sensible approach.